+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `shape = "rcollapse"` and `shape = "rbind"`
 * `glance_custom()` can drop GOF by assigning `NA`: https://stackoverflow.com/questions/75215355/assigning-different-coefficient-names-and-goodness-of-fit-statistics-to-differen
+* When a `statistic` is not available, `modelsummary` prints an empty cell instead of returning an error.
 
 Bug fixes: